Added missing spark submit parameters. These are required for proper working of the yarn-launcher.
